# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hello management API",
    "version": "v1",
    "description": "hello-control REST API. Every route except version, openapi.json and auth/login requires the hello_session cookie or an `Authorization: Bearer <token>` header. Secrets (device SIP secrets, API tokens, trunk passwords) appear only in the response that creates them; trunk passwords never appear in any response."
  },
  "security": [
    {
      "sessionCookie": []
    },
    {
      "bearerToken": []
    }
  ],
  "paths": {
...
    "/api/v1/trunks": {
      "get": {
        "operationId": "listTrunks",
        "summary": "Every trunk (never its password)",
        "tags": ["trunks"],
        "responses": {
          "200": {
            "description": "Trunks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["items"],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Trunk"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      },
      "post": {
        "operationId": "createTrunk",
        "summary": "Create a trunk",
        "tags": ["trunks"],
        "description": "Validated field by field, then as a whole configuration with the routing engine; a 400 lists every failing field in error.fields. 503 when this hello-control is built without the routing engine.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrunkCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trunk"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
...
    "/api/v1/trunks/{id}": {
      "get": {
        "operationId": "getTrunk",
        "summary": "One trunk (never its password)",
        "tags": ["trunks"],
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Trunk",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trunk"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      },
      "patch": {
        "operationId": "updateTrunk",
        "summary": "Change a trunk",
        "tags": ["trunks"],
        "description": "Validated field by field, then as a whole configuration with the routing engine; a 400 lists every failing field in error.fields. Destinations and sourceCidrs, when given, replace the lists. 503 when this hello-control is built without the routing engine.",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrunkUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trunk"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      },
      "delete": {
        "operationId": "deleteTrunk",
        "summary": "Delete a trunk",
        "tags": ["trunks"],
        "description": "409 while an inbound or outbound route names the trunk. 503 when this hello-control is built without the routing engine.",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
    "/api/v1/routes/outbound": {
      "get": {
        "operationId": "listOutboundRoutes",
        "summary": "Every outbound route in position order",
        "tags": ["routes"],
        "responses": {
          "200": {
            "description": "Routes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["items"],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/OutboundRoute"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      },
      "post": {
        "operationId": "createOutboundRoute",
        "summary": "Append an outbound route",
        "tags": ["routes"],
        "description": "Validated field by field, then as a whole configuration with the routing engine; a 400 lists every failing field in error.fields. 503 when this hello-control is built without the routing engine.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OutboundRouteCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutboundRoute"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
...
    "/api/v1/routes/outbound/{id}": {
      "get": {
        "operationId": "getOutboundRoute",
        "summary": "One outbound route",
        "tags": ["routes"],
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Route",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutboundRoute"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      },
      "patch": {
        "operationId": "updateOutboundRoute",
        "summary": "Change an outbound route",
        "tags": ["routes"],
        "description": "Validated field by field, then as a whole configuration with the routing engine; a 400 lists every failing field in error.fields. 503 when this hello-control is built without the routing engine.",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OutboundRouteUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutboundRoute"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      },
      "delete": {
        "operationId": "deleteOutboundRoute",
        "summary": "Delete an outbound route",
        "tags": ["routes"],
        "description": "503 when this hello-control is built without the routing engine.",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
    "/api/v1/routes/inbound": {
      "get": {
        "operationId": "listInboundRoutes",
        "summary": "Every inbound route in position order",
        "tags": ["routes"],
        "responses": {
          "200": {
            "description": "Routes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["items"],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/InboundRoute"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      },
      "post": {
        "operationId": "createInboundRoute",
        "summary": "Append an inbound route",
        "tags": ["routes"],
        "description": "Validated field by field, then as a whole configuration with the routing engine; a 400 lists every failing field in error.fields. 503 when this hello-control is built without the routing engine.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InboundRouteCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InboundRoute"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
...
    "/api/v1/routes/inbound/{id}": {
      "get": {
        "operationId": "getInboundRoute",
        "summary": "One inbound route",
        "tags": ["routes"],
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Route",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InboundRoute"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      },
      "patch": {
        "operationId": "updateInboundRoute",
        "summary": "Change an inbound route",
        "tags": ["routes"],
        "description": "Validated field by field, then as a whole configuration with the routing engine; a 400 lists every failing field in error.fields. 503 when this hello-control is built without the routing engine.",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InboundRouteUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InboundRoute"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      },
      "delete": {
        "operationId": "deleteInboundRoute",
        "summary": "Delete an inbound route",
        "tags": ["routes"],
        "description": "503 when this hello-control is built without the routing engine.",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "hello_session"
      },
      "bearerToken": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "parameters": {
      "id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer",
          "format": "int64",
          "minimum": 1
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid input",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid credentials",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "Conflict": {
        "description": "A unique value is already taken",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Internal": {
        "description": "Internal error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unavailable": {
        "description": "Live state (Valkey) is unavailable",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
...
      "Trunk": {
        "type": "object",
        "required": [
          "id",
          "hasPassword",
          "createdAt",
          "updatedAt",
          "name",
          "mode",
          "username",
          "realm",
          "fromDomain",
          "registerExpires",
          "optionsInterval",
          "sourceCidrs",
          "maxCalls",
          "defaultCallerId",
          "enabled",
          "destinations"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "pattern": "^[A-Za-z0-9._-]{1,64}$",
            "example": "carrier-primary"
          },
          "mode": {
            "type": "string",
            "enum": ["registration", "ip"]
          },
          "username": {
            "type": "string",
            "maxLength": 128
          },
          "realm": {
            "type": "string",
            "maxLength": 128
          },
          "fromDomain": {
            "type": "string"
          },
          "registerExpires": {
            "type": "integer",
            "minimum": 60,
            "maximum": 86400,
            "description": "Seconds"
          },
          "optionsInterval": {
            "type": "integer",
            "minimum": 5,
            "maximum": 3600,
            "description": "Seconds"
          },
          "sourceCidrs": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "type": "string",
              "example": "203.0.113.0/24"
            }
          },
          "maxCalls": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100000,
            "description": "0 = unlimited"
          },
          "defaultCallerId": {
            "type": "string",
            "pattern": "^(\\+?[0-9]{2,20})?$",
            "example": "+97142000000"
          },
          "enabled": {
            "type": "boolean"
          },
          "destinations": {
            "type": "array",
            "minItems": 1,
            "maxItems": 16,
            "items": {
              "$ref": "#/components/schemas/Destination"
            }
          },
          "hasPassword": {
            "type": "boolean",
            "description": "Whether a password is set; the password itself is never returned."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}